move getResult method into Match class
