+++ apps/src/sites/studio/pages/levels/_match.js
@@ -1,11 +1,7 @@
 /* global appOptions */
 import React from 'react';
-import {registerGetResult} from '@cdo/apps/code-studio/levels/codeStudioLevels';
 import {showDialog} from '@cdo/apps/code-studio/levels/dialogHelper';
-import {
-  MatchAngiGifDialog,
-  MatchErrorDialog
-} from '@cdo/apps/lib/ui/LegacyDialogContents';
+import {MatchAngiGifDialog} from '@cdo/apps/lib/ui/LegacyDialogContents';
 
 import Match, {initMatch} from '@cdo/apps/code-studio/levels/match';
 window.Match = Match;
@@ -19,36 +15,4 @@ $(function() {
 
   const container = $('#level-body').first();
   initMatch(container, true);
-});
-
-registerGetResult(() => {
-  let wrongAnswer = false;
-
-  const elements = $('#slots li');
-
-  const response = [];
-
-  for (let index = 0; index < elements.length; index++) {
-    const originalIndex = elements[index].getAttribute('originalIndex');
-    response.push(originalIndex);
-    if (originalIndex === null) {
-      // nothing dragged in this slot yet
-      wrongAnswer = true;
-
-      $('#xmark_' + index).hide();
-    } else if (originalIndex !== String(index)) {
-      // wrong answer
-      wrongAnswer = true;
-
-      $('#xmark_' + index).show();
-    } else {
-      // correct answer
-      $('#xmark_' + index).hide();
-    }
-  }
-  return {
-    response: response,
-    result: !wrongAnswer,
-    errorDialog: wrongAnswer ? <MatchErrorDialog /> : null
-  };
 });
